Fix 'connection' attribute for Door and Building
When parsing a region, a Door needs to connect to the outside ("_front") and Building needs to connect to the inside ("_interior"). We can get these references from the n/s/e/w direction attributes for the region. Building never had the connection attribute and Door was wrong (pointed to itself).

Also, fix Teleport and Elevator to have an 'address' attribute. It's empty for now since it will need to be specified through some other means. None of the .reg files we have seen have anything other than 20 bytes all set to 32, which is uninitialized.