Region changes!
Depth fixes for the following areas:
 - D'nalsi Island
 - I/5
 - Downtown
 - Popustop
 - Streets
 - Woods

Popustop will require a lot of work and so only the ground floor has been corrected. D'nalsi Island will also need further work as a couple of regions seem to have graphical issues, however they work fine in the db dump. Fixes for these will come in a later PR.

Lighting has been turned off in the D'nalsi Island caves. People now need a flashlight to explore just like in the original!

A choke machine was added to the Woods based on findings from the db dump in Woods_6w.