Fix Docker build
Addresses #437.

NodeSource has reinstated the deprecated installer scripts, but the NodeSource repo doesn't contain a packge for npm, and the CentOS package for npm conflicts with the NodeSource nodejs. Updated to use the nsolid package directly.

Also updated the habibots Dockerfile to use the latest node 18 image. Debian Bookworm transitioned the netcat package from one that installed netcat-openbsd to a "virtual package" which doesn't specify a preferred version. Changed it to specify netcat-openbsd explicitly.